# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Atlas Backend API",
        "description": "API documentation for Atlas Backend - Project Management System",
        "contact": {
            "email": "support@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/pg/projects": {
            "get": {
                "tags": [
                    "Projects"
                ],
                "summary": "Get all projects",
                "operationId": "c2196c0c84bc15b75105199b161dbca2",
                "responses": {
                    "200": {
                        "description": "List of projects with related group/member metadata",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ProjectResource"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Projects"
                ],
                "summary": "Create a new project",
                "operationId": "1cd2b1a507a5b21ac836632b071fb6f9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Nuevo proyecto"
                                    },
                                    "proposal_id": {
                                        "type": "integer",
                                        "example": 8,
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Project created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProjectDetailResource"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/pg/projects/{id}": {
            "get": {
                "tags": [
                    "Projects"
                ],
                "summary": "Get a specific project",
                "operationId": "0e23dd22a8267d1b63872e27b9bb0a21",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Project details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProjectDetailResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Project not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Projects"
                ],
                "summary": "Update a project",
                "operationId": "4e07ed33b36cf089bcce90d1e702902e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "status_id": {
                                        "type": "integer",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Project updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProjectDetailResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Project not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Projects"
                ],
                "summary": "Delete a project",
                "operationId": "181f27ca5f9ca83afb6eafc916c4cc9a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Project deleted successfully"
                    },
                    "404": {
                        "description": "Project not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ProjectResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 12
                    },
                    "title": {
                        "type": "string",
                        "example": "Plataforma IoT"
                    },
                    "status": {
                        "type": "string",
                        "example": "Activo"
                    },
                    "member_names": {
                        "type": "string",
                        "example": "Laura Proposer, Juan Developer"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "ProjectDetailResource": {
                "description": "Minimal project representation; related entities available via dedicated endpoints.",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 12
                    },
                    "title": {
                        "type": "string",
                        "example": "Plataforma IoT"
                    },
                    "status": {
                        "type": "string",
                        "example": "Activo"
                    },
                    "proposal_id": {
                        "type": "integer",
                        "example": 5,
                        "nullable": true
                    },
                    "thematic_line_name": {
                        "type": "string",
                        "example": "Inteligencia Artificial",
                        "nullable": true
                    },
                    "member_names": {
                        "type": "string",
                        "example": "Laura Proposer, Juan Developer"
                    },
                    "staff_names": {
                        "type": "string",
                        "example": "Ana Directora, Luis Asistente"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}